requirements.txt
appdirs==1.4.3
fpdf==1.7.2
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.2.0
PyYAML==3.12
six==1.10.0
